style: 💄 refactored the project structure
