+++ src/pages/DuplicateBeneficairiesPage/components/DuplicateBeneficiariesContainer/components/CustomTable/CustomTable.module.css
@@ -13,7 +13,6 @@
 
 .table-body {
     display: block;
-    min-height: 10vh;
     max-height: 50vh;
     overflow-y: auto;
     overflow-x: hidden;
@@ -23,4 +22,3 @@
     width: 100%;
     table-layout: fixed;
 }
-
