+++ src/pages/DuplicateBeneficairiesPage/components/DuplicateBeneficiariesContainer/components/TableContainer/TableContainer.module.css
@@ -1,0 +1,4 @@
+.center {
+    margin: 0 auto;
+    text-align: center;
+}
