+++ src/pages/DuplicateBeneficairiesPage/index.tsx
@@ -1,8 +1,7 @@
 import React, { useState } from "react";
-import CustomTableContainer from "../../components/CustomTableContainer";
-import DataSelectionContainer from "../../components/DataSelectionContainer";
-import { SelectionDimension } from "../../components/DataSelectionContainer/interfaces";
-import classes from "./DuplicateBeneficiariesPage.module.css";
+import CustomTableContainer from "./components/DuplicateBeneficiariesContainer";
+import DataSelectionContainer from "./components/DataSelectionContainer";
+import { SelectionDimension } from "./components/DataSelectionContainer/interfaces";
 
 export function DuplicateBeneficiariesPage(): React.ReactElement {
   const [dataSelection, setDataSelection] = useState<
